requirements.txt
streamlit>=1.24.0
networkx>=3.1
pyvis>=0.3.2
pandas>=2.0.0
numpy>=1.24.0
python-arango>=7.5.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
uvicorn
requests
scikit-learn
pydantic 